+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enhance `MixedNamesRule` with symbol table information.
 - Exclude class constructors in `ConstructorCreateRule`.
 - Handle `else` exception handler in `SwallowedExceptionsRule`.
+- Handle global variables separately in `VariableNameRule`.
 
 ### Fixed